refactor(api): extract shared captcha validation module #1134
